BeeNeXT version 3.0.0
- สร้างใหม่ โปรโตคอลใหม่ สามารถรับส่ง byte ตามจำนวนที่ต้องการได้
และมีระบบ checksum CRC16
- รองรับ UNO, MEGA, NANO, STM32Bluepill, ESP8266, ESP32

+++ examples/03.TwoMCUs_communication/ArduionMEGA_code/ArduionMEGA_code.ino
@@ -15,8 +15,8 @@ void setup() {
 
   static SoftTimer timer;
   timer.setInterval(1000,[](){
-    BeeNeXT.send("TEMP" , random(100.0)/10.0 );    // ตัวอย่างส่งแบบ key & value จาก MCU ปัจจุบันไปอีก MCU
-    BeeNeXT.send("HUMID", random(100.0)/10.0 );    // ตัวอย่างส่งแบบ key & value จาก MCU ปัจจุบันไปอีก MCU
+    BeeNeXT.send("TEMP" , (float) random(100.0)/10.0 );    // ตัวอย่างส่งแบบ key & value จาก MCU ปัจจุบันไปอีก MCU
+    BeeNeXT.send("HUMID", (float) random(100.0)/10.0 );    // ตัวอย่างส่งแบบ key & value จาก MCU ปัจจุบันไปอีก MCU
   });
 
 }
